Remove Old Entities (With a Bit of Work on Current User Story) (#10)
* Complete Database Overhaul

Removed the a majority of the old database, including unit tests. Implemeneted corrections to the member and request entities

* Set Status to Default When Created

I don't want the rider to choose the status of their ride. It has been changed so it defaults to "In Progress"

* Small Changes to Names and Strings

* Removed Cancel Link that Breaks Request

* Fixing A Change I Made

Apparently rewording this broke the site.

* Added Link to Cancel Request

Adding a custom function to controller in order to use in view requires adding the function to the routes.